+++ kubernetes/chart/zulip/CHANGELOG.md
@@ -1,3 +1,14 @@
+## [0.11.4] - 2025-10-23
+
+- Update to Zulip Server 11.4
+
+## [0.11.3] - 2025-10-22
+
+- Update to Zulip Server 11.3
+- Support valueFrom in both `SECRETS_` and `SETTING_` values
+- Support `PROXY_ALLOW_*` settings for outgoing proxy rules
+- Support `DB_NAME` / `DB_USER` settings.
+
 ## [0.11.2] - 2025-09-16
 
 - Update to Zulip Server 11.2
@@ -130,7 +141,6 @@
 ## [0.3.0] - 2022-04-21
 
 - Update dependencies:
-
   - Helm charts:
 
     | Repository                         | Name       | Version |